refactor(codegen): split masked-window load helpers into expr/masked_window.rs
The masked-window fact consult + the two load-emission tiers pushed
expr/index_get.rs past the 2000-line lint cap (check_file_size.sh); move
them to a topical sibling module. Pure move plus factoring the shared
raw in-window load emission; no behavior change (benchmarks and the
edge-case suite are byte-identical before/after).
